refactor(config): extract `findConfigToken` helper and simplify token resolution
Verified: precommit-verify ✕
