+++ packages/core/src/config/index.ts
@@ -1,3 +1,4 @@
-export { Config } from './decorator';
-export { injectConfig } from './inject';
-export type { ConfigClass } from './types';
+export {findConfigToken} from "./token";
+export {Config} from './decorator';
+export {injectConfig} from './inject';
+export type {ConfigClass} from './types';
