+++ packages/core/src/internal/container.ts
@@ -1,4 +1,5 @@
 import { Container } from '@needle-di/core';
+import {findConfigToken} from "../config";
 
 type Class<T> = new (...args: never[]) => T;
 
@@ -10,18 +11,6 @@ export type ResolverHandle = {
   readonly get: <T extends object>(cls: Class<T>) => T;
 };
 
-type AnyClass = new (...args: never[]) => unknown;
-
-const findTokenOwner = (cls: AnyClass): AnyClass | null => {
-  let current: AnyClass | null = cls;
-  while (current && current !== Function.prototype) {
-    if ('Token' in current) {
-      return (current as { Token: AnyClass }).Token;
-    }
-    current = Object.getPrototypeOf(current) as AnyClass | null;
-  }
-  return null;
-};
 
 // Controllers / Service / Adapter は `@Controller` または `@injectable()` decorator により
 // needle-di が `container.get(cls)` 時に auto-bind する。明示的な bind は持たない (spec §4.10)。
@@ -29,8 +18,8 @@ export const createContainer = (options: CreateContainerOptions = {}): ResolverH
   const container = new Container();
 
   for (const configClass of options.configs ?? []) {
-    const token = findTokenOwner(configClass);
-    if (token && token !== configClass) {
+    const token = findConfigToken(configClass);
+    if (token) {
       container.bind(configClass);
       container.bind({ provide: token, useExisting: configClass });
     }
